% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZCoefficientFunctions.R
\docType{data}
\name{cancer_mutations}
\alias{cancer_mutations}
\title{Gene mutations in cancers}
\format{A data frame representing coningency table.}
\source{
The \code{cancer_mutations} is taken from the following article 
  \url{https://www.nature.com/articles/nature12634}.
}
\usage{
data(cancer_mutations)
}
\description{
The \code{cancer_mutations} is a matrix (contingency table) representing 
different gene mutations in different cancer types.
}
\details{
Sample dataset was taken from (Kandoth et al., (2013) Mutational
landscape and significance across 12 major cancer types. Nature, 502,
333–339.) and consist of information related to mutated genes (with
point mutations and small insertions/deletions) from 3281 tumours across 12
tumour types, groupedd into 11 categories.
Rows correspond to different tumore types and columns to different genes.
}
\seealso{
Other cancer mutations: \code{\link{cancer_mutations_gene_groups}}
}
\keyword{datasets}
