% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZCoefficientFunctions.R
\docType{data}
\name{cancer_mutations_gene_groups}
\alias{cancer_mutations_gene_groups}
\title{Gene groups for gene mutations in cancers}
\format{A factor vector}
\source{
Information regarding biochemical processes in which genes are taking
  part is taken from the following article/database/website: \url{to be supplemented}.
}
\usage{
data(cancer_mutations_gene_groups)
}
\description{
The \code{cancer_mutations_gene_groups} is a vector of factors specifying gene
groups for different gene mutations in different cancer types stored in
\code{cancer_mutations}.
}
\details{
The factor vector \code{cancer_mutations_gene_groups} defines a biochemical
process in which particular gene is taking part.
}
\seealso{
Other cancer mutations: \code{\link{cancer_mutations}}
}
\keyword{datasets}
