% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZCoefficientFunctions.R
\name{z_coefficient_matrix}
\alias{z_coefficient_matrix}
\title{Z coefficient evaluation function in a matrix version}
\usage{
z_coefficient_matrix(m, col_groups = NULL, row_groups = NULL)
}
\arguments{
\item{m}{A matrix. It can be a contingency table with counts or a table with
probabilities. It is expected that there are no negative entries in a
matrix and that mariginal sums are positive for all columns and rows.}

\item{col_groups, row_groups}{A vector or factor giving the grouping, with one
element per column (row) of \code{m}. Missing values will be treated as
another group and a warning will be given. When a row (column) grouping is
specified it is assumed that mariginal sums for each group is positive.}
}
\value{
A matrix of Z coefficients.
}
\description{
The function \code{z_coefficient_matrix} returns a matrix of Z coefficients.
}
\details{
Z coefficient \code{Z[A_i,B_j]} for a matrix \code{m} is defined by the
following formula: \deqn{Z(A_i,B_j)^2 = 1 -
[P(B_j)*P(A_i|B_j)/P(A_i)*P(A_i^c|B_j)/P(A_i^c) + 
P(B_j^c)*P(A_i|B_j^c)/P(A_i)*P(A_i^c|B_j^c)/P(A_i^c)],} where \code{A_i}
corresponds to the row \code{i} and \code{B_j} corresponds to the column
\code{j}.

The function \code{z_coefficient_matrix} calculates a matrix, where each entry
\code{m[i,j]} of a matrix \code{m} is the Z coefficient \code{Z(A_i,B_j)}.
}
\examples{
#load gene mutations in cancer dataset together with gene groups
data("cancer_mutations")
data("cancer_mutations_gene_groups")

#calculating Z coefficients in a matrix form, also using gene groups
#head used to show only top 6 rows
head(z_coefficient_matrix(cancer_mutations))
head(z_coefficient_matrix(cancer_mutations, col_groups = cancer_mutations_gene_groups))
}
\seealso{
Other Z coefficient: \code{\link{z_coefficient_ranks}},
  \code{\link{z_coefficient}}
}
\concept{
Z,coefficient,association
}
